chore : create guard, strategy directory
