+++ src/auth/strategy/local.strategy.ts
@@ -1,0 +1,20 @@
+import {Strategy} from 'passport-local';
+import {PassportStrategy} from '@nestjs/passport';
+import {BadRequestException, Injectable, UnauthorizedException} from '@nestjs/common';
+import {AuthService} from '../auth.service';
+import {Err} from 'src/common/error';
+
+@Injectable()
+export class LocalStrategy extends PassportStrategy(Strategy) {
+  constructor(private authService: AuthService) {
+    super({usernameField: 'email', passwordField: 'password'});
+  }
+
+  async validate(email: string, password: string): Promise<any> {
+    const user = await this.authService.validateUser(email, password);
+    if (!user) {
+      throw new BadRequestException(Err.USER.NOT_FOUND);
+    }
+    return user;
+  }
+}
